requirements.txt
Flask==2.3.0
psycopg2-binary==2.9.7
requests==2.31.0

